Moved test files to corresponding position according to received commentary
